refactor: move tests to reporter repo
