+++ reporter/.nycrc.json
@@ -1,0 +1,13 @@
+{
+    "reporter": [
+        "text",
+        "lcov",
+        "json",
+        "text-summary"
+    ],
+    "include": [
+        "lib/**/*.js"
+    ],
+    "report-dir": "./.coverage",
+    "temp-dir": ".nyc_output"
+}
