+++ reporter/package.json
@@ -7,7 +7,11 @@
     "lib": "lib"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "npm run test:lint && npm run test:suite ",
+    "test:lint": "eslint .",
+    "test:suite": "npm run test:unit",
+    "test:unit": "nyc mocha \"test/unit/**/*.test.js\"",
+    "coverage": "codecov"
   },
   "author": "elit-altum",
   "license": "Apache-2.0",
@@ -18,5 +22,20 @@
     "nanoid": "3.1.20",
     "pidusage": "2.0.21",
     "socket.io-client": "4.1.2"
+  },
+  "devDependencies": {
+    "chai": "^4.3.4",
+    "codecov": "^3.8.2",
+    "eslint": "^7.30.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-promise": "^5.1.0",
+    "mocha": "^8.4.0",
+    "mock-require": "^3.0.3",
+    "nyc": "^15.1.0",
+    "prettier": "^2.3.0",
+    "sinon": "^11.1.1"
   }
 }
